Question4

+++ AreaofaCircle.java
@@ -3,7 +3,7 @@ import java.util.*;
 class AreaofaCircle{
 public static void main(String args[]){
 Scanner kb=new Scanner(System.in);
-int radius=kb.nextInt();
+double radius=kb.nextDouble();
 
 double area=3.14*radius*radius;
 System.out.println("Area of Circle is "+ area);
